pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.9"

dependencies = [
    "cartesia>=2.0.17",
    "elevenlabs>=2.23.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "livekit>=1.0.19",
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-api>=1.0.7",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai~=1.0",
    "livekit-plugins-cartesia~=1.0",
    "livekit-plugins-deepgram~=1.0",
    "pydub>=0.25.1",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
